APIGOV-27199 Regenerate api-server models

+++ pkg/apic/apiserver/models/management/v1alpha1/model_traceability_agent_spec_config_sampling.go
@@ -11,8 +11,9 @@ package management
 
 // TraceabilityAgentSpecConfigSampling How the agent will sample the transaction data found during collection
 type TraceabilityAgentSpecConfigSampling struct {
-	// The number of transactions, out of 100, that the agent will save the full transaction details for, valid values are 0 to 50
-	Percentage int32 `json:"percentage"`
+	// The percentage of transactions that the agent will save the full transaction details for. Valid values(decimals included) are 0 to 10
+	// GENERATE: The following code has been modified after code generation
+	Percentage float64 `json:"percentage"`
 	// When set to true will send transactional data for every errored transaction for Business and Consumer Insights
 	AllErrors bool `json:"allErrors"`
 }
